Added Arc'd Mutex for individual bot data.
This will ensure that we don't block the requests related to other
conversations, since the lock on global bot data is released immediately
after getting the data. Ideally, we should queue another Future following
the Response, but I can't figure it out just yet.

All of this is to make sure that the TCP connections made by the bot and
the handler implemented by the client don't affect the response time
of our server itself.